+++ packages/core/tests/fixtures/basic-app/src/app.ts
@@ -5,12 +5,5 @@ export const app = express();
 app.use(express.json());
 
 export const boot = async () => {
-  return await createApp(app, {
-    modules: 'src/modules/*',
-    prefix: '/api',
-    aliases: {
-      '@config': './src/config',
-      '@middleware': './src/middleware'
-    }
-  });
+  return await createApp(app);
 };